requirements.txt
Django==1.9.3
django-tastypie==0.13.3
psycopg2==2.6.1
python-dateutil==2.5.0
python-mimeparse==1.5.1
six==1.10.0
